ADD: Función completa para borrar un documento asociado a una solicitud

+++ server/services/documents.service.js
@@ -1,11 +1,24 @@
 import Document from "../models/documentModel.js"
+import fs from 'fs/promises'
+import path from 'path'
 
 // Borrar un documento asociado a una solicitud
 export const deleteDocumentService = async (id) => {
     try {
+        const document = await Document.findByPk(id)
+        if (!document) {
+            throw new Error("No se encontró el documento")
+        }
+
+        // Obtener ruta absoluta del documento
+        const documentPath = path.resolve(document.ruta)
+
+        // Borrar documento del servidor
+        await fs.unlink(documentPath)
+
+        // Borrar registro de la base de datos
         await Document.destroy({ where: { id } })
     } catch (error) {
-        console.log(error)
-        return null
+        throw new Error("No se pudo borrar el documento")
     }
 }
